+++ src/core/compiler/backend/railshot/amd64/stack_test.go
@@ -165,7 +165,7 @@ func TestAssignPinnedLocalsUsesLocalDefs(t *testing.T) {
 		localType: []machineType{mtI32, mtF64, mtI32},
 		sc:        &scratch{},
 	}
-	f.assignPinnedLocals([]int64{1, 10, 5}, nil, nil, pinnedLocalRegs, baseFPPins, false)
+	f.assignPinnedLocals([]uint32{1, 10, 5}, nil, nil, pinnedLocalRegs, baseFPPins, false)
 
 	r, isFloat, ok := f.pinReg(1)
 	if !ok || !isFloat || r != pinnedFLocalRegs[0] {
